Added Springs and Stones
Added the sprites for the spring as well as the stones. Blue stone and spring collapsed are currently not used or functional properly.

Added functions to create and collect the stones as well as create the springs on every 5th ledge as well as a collison function for the spring-