Added Stones with Platform Collision Check
Added the stone-platfrom collision check in the update function. everytime a stone collides with the platfrom the stone gets moved.

I also changed the regenerating function of the stones so that new stones get generated when the player passes the 4th last stone not only the last. thought that might improve the game flow a bit.