Changed Spring Function
Added the sping generating function to the create function that only 1 spring gets generated. the update function checks if the spring is out of sight and moves it in sight again on spring 6 or 7. Also changed to spring-player collision function so that the player only jumps when touching down on the spring.